face-recog/requirements.txt
boto3~=1.29.6
dlib
numpy
face_recognition
opencv-python
ffmpeg
awscli
smart_open
python-dotenv~=1.0.0
requests~=2.31.0